WIP: animate boids based on potential fields
I'm building an ... arena I guess you could say ... that the boids will
"fly" around in: the middle of the arena is shaped like a pyramid with
rounded corners and the outside is like a moat with a round bottom. The
effect should be that the boids always stay in the moat. I'm also going
to "tip" the arena for each boid so that they all "float" around the
moat in the same direction.