WIP: fix tests while working on flockVector
Since I moved the velocity cap to aggregatedAdjustments, rather than
capping each individual adjustment (and unfairly weighting them) I had
to find out by what proportion the aggregated adjustment went over the
distance, and then subtract that proportion from X & Y;

e.g. if the velocity is 10 and the adjustment is 20 away from the boid's
point, then the adjustment has to be cut in half, so the X adjustment
and Y adjustment are both cut in half.