Fix & Test combinePoints, added pointsRelativeTo
pointsRelativeTo takes a start point and a bunch of other points and
converts the other points to points that are relative to the start
point: e.g. if start point is 10,10 and an "other" point is 12,8 then
the "other" point becomes 2,-2, i.e. what 12,8 is relative to 10,10.

This allows us to get a bunch of target points and figure out what
distances they represent (x and y distances) in relation to a reference
point: i.e. if I want to move from point p to point q, what is p.x - q.x
and p.y - q.y.