WIP: Test applyMovementPipeline (plus logging)
I'm testing out the applyMovementPipeline function, but I'm not done,
and I have some other logging functions I threw in while furiously
debugging; I realized after a while that it would smart to make sure
boids.js is tested (because it will undoubtedly turn out to be broken,
and it was) before I waste time trying to debug other code that's
already tested pretty well.

I either need to do the math and figure out what the two current
movement functions should do to a point based on a couple of neighbours
and a canvas position, or mock out those functions. I think it would be
a tremendous exercise to actually figure out what the flock and arena
movement functions will combine to do to a few different points with a
few different setups. Is it redundant? Yes, in a way, but I'd still like
to have it just so I can sleep better.