Moved animate function to boids.html
The animate function now sets up a "world" data structure (note that I'm
conspicuously not using the term "object") to pass to the animateBoids
function. Most functions now take the "world" data to operate on so that
movement functions in the movement pipeline (and, in fact, most
functions) can use the same interface.

Added a "Log?" checkbox to turn off logging: innerHTML changes were
taking a TON of CPU.

Log() will be commented on in a later commit.

This is the first of (I suspect) a bunch of commits as I've done a TON
of work, but wanted to give some good commit comments when I could give
them some thought.