Tons of debugging changes ...
In order to get flocking to work (albeit poorly) I had to do a ton of
changes.

I've moved the "animate" function to boids.html where it now sets up a
common interface as the "world" data structure. (see earlier commit)

Functions will now "change the world" using the changeWorldBoids
function to update between three fields in the world associative array:
boid, boids and movedBoids.

"boid" is the current boid from anyone's perspective;
"boids" are the boids that have not been operated on;
"movedBoids" are boids that have been operated on.

Typically we grab the next boid in boids, set "boid" to it, operate on
boid, then when we're done we transfer it to movedBoids (possibly with
shiftWorld).

combinePoints now returns the boid's location if no changes are made
(e.g. because no boids are near)

I've added a bunch of debuging "toString" functions (and tests) to make
logging and debugging MUCH easier.