WIP: weight movement functions to share velocity
Currently the flocking "repel" behaviour and the "arena" (or river,
moat, etc.) behaviour conflict and the velocity limit applies to each
equally; since the "arena" behaviour always move the same distance and
the "repel" behaviour varies, the "arean" behaviour dominates: this is
why we see the boids line up along the current instead of across it.

My plan is to weight the movement functions so that flocking has more
pull than the arena. If the "repel" behaviour only pushes 1 distance,
where arena pushes 5, but "repel" is weighted at .8, then the effects
should be more even. The boids will still probably line up, but this can
be tweaked.

If the combined movement functions exceed the velocity, then the allowed
movement should be shared by the movement functions according to their
weights:

i.e. if A moves 10 and has a weight of .8, and B moves 5 and has a
weight of .2, but the velocity is 9, then we can only move 60% of the
distance; A gets .8 of the velocity and B gets .2; i.e. A gets 7.2 and B
gets 1.8.

I don't like the feel of this because A moved 100% more than B, but gets
400% more movement in the end; it kind of feels like A should get 800%
more movement since it had four times the weight plus twice the
movement.

Another option is to pass the velocity to each movement function and
have it adjust it's movement according to it's share of the velocity.